sem6/AI/zad5/requirements.txt
torch
torchvision
facenet-pytorch
numpy
Pillow
requests
tqdm
deeplake
scikit-learn
pandas
gdown